got things working completely using the new boost.multiprecision.complex!
but, many tests fail with my hand-roll now.  sad day.
